feat: create widget and detail activity
I created a widget that can be placed on the homescreen that displays
the most recently created image and provides a way to launch the
application. I also created a detail activity that launches when a
thumbnail in the gallery is clicked on to display a higher resolution
version of the image, view the artists, and allow deletion from the
database.